Give jobs unique names (#43)
The previous system that used the same name for the same type of job,
e.g. `Lint code`, made it very easy to configure branch protection
rules. Since the jobs shared the same name across file extensions, a
single required status check covered all jobs.

When allowing Renovate to merge its own pull requests, we discovered
an issue with this approach. When a pull request was opened, Renovate
would only wait for the first required status checks to be skipped.
GitHub counts skipped checks as successful, so from the perspective of
Renovate all tests at that moment had succeeded. It wasn't aware of
checks with the same name starting a little bit later.

This timing issue cannot be fixed with Renovate's configuration. So we
are switching to unique job names that can be required individually as
status checks. This enables us to force Renovate to wait until all
languages have determined whether its tests must run.